requirements_training.txt
# Requirements for EMONET finetuning
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.1.0
scikit-image>=0.15.0
face-alignment>=1.3.0
numpy>=1.19.0
matplotlib>=3.3.0
tqdm>=4.60.0
Pillow>=8.0.0
pathlib2>=2.3.0
